requirements.txt
gspread==6.1.4
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
Jinja2==3.1.5
